+++ doc/Sphinx/releases.rst
@@ -28,6 +28,7 @@ Changes made in the repository (not released)
   * normalized laser frequency can be different from 1
 
 * Particles can be imported from a file
+* Some :doc:`profiles` can be imported from a file
 * Coulomb logarithm may be multiplied by a constant factor
 * Happi: